PubMatic RTD Provider: Move bidderOptimisation to one file
